[refactor] #48 presentation layer  model , type은 공통으로 사용되는게 아니면 contract package내부에 선언  navigation은 각 메인 스크린 navigation package에 선언
