+++ app/src/main/java/org/android/bbangzip/presentation/ui/timer/TimerViewModel.kt
@@ -8,10 +8,11 @@ import kotlinx.coroutines.delay
 import org.android.bbangzip.domain.usecase.CompleteTimerUseCase
 import org.android.bbangzip.domain.usecase.FetchBreadListUseCase
 import org.android.bbangzip.domain.usecase.UpdateTodayBreadCountUseCase
-import org.android.bbangzip.presentation.common.model.TimerStatus
+import org.android.bbangzip.presentation.ui.timer.contract.model.TimerStatus
 import org.android.bbangzip.presentation.ui.timer.lifecycle.TimerLifecycleManager
 import org.android.bbangzip.presentation.common.base.BaseViewModel
 import org.android.bbangzip.presentation.common.util.constant.TimerConstants
+import org.android.bbangzip.presentation.ui.timer.contract.TimerContract
 import timber.log.Timber
 import javax.inject.Inject
