+++ app/src/main/java/org/android/bbangzip/presentation/ui/timer/lifecycle/TimerLifecycleManager.kt
@@ -12,7 +12,7 @@ import dagger.assisted.AssistedInject
 import dagger.hilt.android.qualifiers.ApplicationContext
 import org.android.bbangzip.presentation.common.observer.AppLifecycleObserver
 import org.android.bbangzip.presentation.common.observer.ScreenStateReceiver
-import org.android.bbangzip.presentation.ui.timer.TimerContract
+import org.android.bbangzip.presentation.ui.timer.contract.TimerContract
 import timber.log.Timber
 
 class TimerLifecycleManager
