+++ app/src/main/java/org/android/bbangzip/presentation/ui/timer/navigation/TimerTodo.kt
@@ -1,0 +1,6 @@
+package org.android.bbangzip.presentation.ui.timer.navigation
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TimerTodo(val timeOptionIndex: Int = 0)
