+++ app/src/main/java/org/android/bbangzip/presentation/ui/timer/todo/TimerTodoViewModel.kt
@@ -6,19 +6,19 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import org.android.bbangzip.presentation.common.model.Category
 import org.android.bbangzip.presentation.common.model.ListItem
 import org.android.bbangzip.presentation.common.model.Todo
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoEvent
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoReduce
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoReduce.ClearAddTodoState
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoReduce.UpdateAddTodoBottomSheetState
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoReduce.UpdateCategoriesAndFlatList
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoReduce.UpdateFlatList
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoReduce.UpdateSelectedCategory
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoReduce.UpdateSelectedStartTime
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoReduce.UpdateTimePickerBottomSheetState
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoReduce.UpdateTodoText
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoSideEffect
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoSideEffect.NavigateToTimer
-import org.android.bbangzip.presentation.ui.timer.todo.TimerTodoContract.TimerTodoState
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoEvent
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoReduce
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoReduce.ClearAddTodoState
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoReduce.UpdateAddTodoBottomSheetState
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoReduce.UpdateCategoriesAndFlatList
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoReduce.UpdateFlatList
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoReduce.UpdateSelectedCategory
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoReduce.UpdateSelectedStartTime
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoReduce.UpdateTimePickerBottomSheetState
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoReduce.UpdateTodoText
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoSideEffect
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoSideEffect.NavigateToTimer
+import org.android.bbangzip.presentation.ui.timer.todo.contract.TimerTodoContract.TimerTodoState
 import org.android.bbangzip.presentation.common.base.BaseViewModel
 import java.time.LocalTime
 import javax.inject.Inject
