+++ app/src/main/java/org/android/bbangzip/presentation/ui/timer/todo/contract/TimerTodoContract.kt
@@ -1,10 +1,10 @@
-package org.android.bbangzip.presentation.ui.timer.todo
+package org.android.bbangzip.presentation.ui.timer.todo.contract
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import org.android.bbangzip.presentation.common.base.BaseContract
 import org.android.bbangzip.presentation.common.model.Category
 import org.android.bbangzip.presentation.common.model.ListItem
-import org.android.bbangzip.presentation.common.base.BaseContract
 import java.time.LocalDate
 import java.time.LocalTime
